Refactored PrivateRoute into a common component
