+++ src/components/common/PrivateRoute/PrivateRoute.tsx
@@ -1,13 +1,13 @@
 import React from 'react';
 import { Route, Redirect } from 'react-router-dom';
-import { getToken } from './tokenHelpers';
+import { getToken } from '../../../utils/tokenHelpers';
 
 interface PrivateRouteProps {
   path: string;
   component: React.ComponentType;
 }
 
-export const PrivateRoute = ({
+const PrivateRoute = ({
   component: Component,
   ...props
 }: PrivateRouteProps): React.ReactElement => {
@@ -21,3 +21,5 @@ export const PrivateRoute = ({
     />
   );
 };
+
+export default PrivateRoute;
