+++ src/components/common/PrivateRoute/index.ts
@@ -1,0 +1,1 @@
+export { default as PrivateRoute } from './PrivateRoute';
