+++ src/components/common/index.ts
@@ -1,3 +1,4 @@
 export { Modal } from './Modal';
 export { Header } from './Header';
+export { PrivateRoute } from './PrivateRoute';
 export { AccessControl } from './AccessControl';
