+++ src/utils/index.ts
@@ -1,4 +1,3 @@
-export { PrivateRoute } from './PrivateRoute';
 export { getToken, setToken, clearToken } from './tokenHelpers';
 export { validateSignup } from './formValidation';
 export { formatUTCToLocalTimestring } from './time';
